Preliminary Report
This is really rushed and unfinished, I basically just combined your
files and started to edit the first one. Please let me know if this
document actually combined correctly, because I wasn't working on
Microsoft Word. We should meet at some point to go over this before we
start writing the final report. That said, the graphs all look really
nice and the explanations are clear, it's just not yet combined in a
coherent way. at all. (also i removed the part where we show the missing
data... i think it would be better just to show the missing data on a
simple graph since the -999 values aren't actually representative of
anything to do with PM2.5)